requirements.txt
fastapi
pydantic
pymongo[srv]
uvicorn
gunicorn
python-dotenv
motor
requests
python-multipart